requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sentence-transformers==2.6.1
faiss-cpu==1.8.0
transformers==4.41.2
numpy==1.26.4
pandas==2.2.2
